refactor: shared calendar-date helpers move into @stll/time
isIsoDateString, parseIsoDateLocal and addDays were duplicated in both apps; the package that already owned DAY_IN_MS owns them now. The API keeps its decision-date bounds and UTC cursor helpers, built on the shared parts.
