+++ apps/web/src/routes/_protected.workspaces/$workspaceId/-components/activity/activity-panel.logic.ts
@@ -1,5 +1,5 @@
 import type { TranslationKey } from "@/i18n/types";
-import { addDays, parseIsoDateLocal } from "@/lib/dates";
+import { addDays, parseIsoDateLocal } from "@stll/time";
 import type { MatterActivityItem } from "@/lib/workspaces/queries";
 
 /**
