+++ apps/web/src/routes/_protected.workspaces/$workspaceId/-components/billing/timesheet-week-view.tsx
@@ -5,7 +5,7 @@ import { cn } from "@stll/ui/utils";
 
 import { getFormattingLocale } from "@/i18n/i18n-store";
 import { normalizeOptionalArray } from "@/lib/arrays";
-import { addDays } from "@/lib/dates";
+import { addDays } from "@stll/time";
 import { timeEntriesOptions } from "@/lib/workspaces/queries/time-entries";
 import { formatCurrencyCompact } from "@/routes/_protected.workspaces/$workspaceId/-components/billing/format-currency";
 import {
